pyproject.toml
[tool.poetry]
name = "rain-alert"
version = "0.1.0"
description = "Python script that sends an email alert if the forecast for the next 12 hours is rain."
authors = ["Igor Ferreira <iferreira.n01@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
